+++ HISTORY.md
@@ -17,6 +17,9 @@ de.otto.edison.discovery to de.otto.edison.registry.
 is to provide cache statistics as HTML and/or JSON, it makes no sense to deactivate the controller.
 * **[edison-jobs-mongo]** This library was combined with edison-jobs. Adding and configuring both edison-jobs
 and edison-mongo will use MongoJobRepository as implementation of the JobRepository interface.
+* **[edison-serviceregistry-client]** Contents of this lib was moved to edison-core. The DiscoveryClient was renamed
+to RegistryClient.
+* **[edison-serviceregistry-client]** Properties edison.servicediscovery.* renamed to edison.serviceregistry.*
 
 **Bugfixes:**
 * Fixed broken link from job messages to /jobdefinitions/<jobType>. JobDefinitionService.getJobDefition(jobType)